+++ lib/src/widgets/xen_emojify.dart
@@ -17,7 +17,7 @@ class XenEmojify extends StatefulWidget {
     this.placeholderWidget = const Icon(Icons.add),
     this.initialEmoji,
     this.lottieSource = LottieSources.network,
-    this.displayLabel = true,
+    this.showLabel = true,
     this.selectedEmojiSize = 25,
     this.selectedEmojiTextStyle = const TextStyle(
       color: Colors.black,
@@ -38,7 +38,7 @@ class XenEmojify extends StatefulWidget {
   /// If true, the name of the selected emoji will be displayed.
   ///
   /// Defaults to true.
-  final bool displayLabel;
+  final bool showLabel;
 
   /// The size of the selected emoji.
   ///
@@ -149,7 +149,7 @@ class _XenEmojifyState extends State<XenEmojify> with TickerProviderStateMixin {
               size: Size(widget.selectedEmojiSize, widget.selectedEmojiSize),
             ),
           ),
-          if (widget.displayLabel) _labelBuilder()
+          if (widget.showLabel) _labelBuilder()
         ],
       );
     }