+++ example/lib/main.dart
@@ -42,44 +42,61 @@ class _XenEmojifyExampleState extends State<XenEmojifyExample> {
 
   @override
   Widget build(BuildContext context) {
-    final lottie = [
+    final lotties = [
       const XenEmoji(
         'https://fonts.gstatic.com/s/e/notoemoji/latest/1f60d/lottie.json',
-        label: 'heart eyes',
+        label: 'Heart Eyes',
         lottieID: '1f60d',
       ),
       const XenEmoji(
-        'https://lottie.host/e9e0b847-de55-4537-8406-d1a9a2be02c7/yMBPVCKSru.json',
-        label: 'laughing',
-        lottieID: '1f606',
+        'https://fonts.gstatic.com/s/e/notoemoji/latest/1f525/lottie.json',
+        label: 'Fire',
+        lottieID: '1f525',
       ),
       const XenEmoji(
-        'https://lottie.host/f7d6b7cf-58f5-4e5e-8a17-8b6087f2e442/A3QjXpmhk9.json',
-        label: 'sunglasses',
-        lottieID: '1f60e',
+        'https://fonts.gstatic.com/s/e/notoemoji/latest/1f4a9/lottie.json',
+        label: 'Poop',
+        lottieID: '1f4a9',
       ),
       const XenEmoji(
-        'https://lottie.host/f47215b5-db87-4c46-82f7-874cb904d7a1/UCA1WVixmM.json',
-        label: 'indian smirk',
+        'https://fonts.gstatic.com/s/e/notoemoji/latest/1f984/lottie.json',
+        label: 'Unicorn',
         lottieID: '1f60f',
       ),
       const XenEmoji(
-        'https://lottie.host/34bf46c9-be14-4e81-a1ba-49e7e64a0d47/g7iC1sdHC2.json',
-        label: 'joy',
+        'https://fonts.gstatic.com/s/e/notoemoji/latest/1faa9/lottie.json',
+        label: 'Mirror Ball',
         lottieID: '1f602',
       ),
       const XenEmoji(
         'https://fonts.gstatic.com/s/e/notoemoji/latest/1f603/lottie.json',
-        label: 'smile',
+        label: 'Smile',
+        lottieID: '1f603',
+      ),
+      const XenEmoji(
+        'https://fonts.gstatic.com/s/e/notoemoji/latest/26a1/lottie.json',
+        label: 'Zap',
+        lottieID: '1f603',
+      ),
+      const XenEmoji(
+        'https://fonts.gstatic.com/s/e/notoemoji/latest/1f47d/lottie.json',
+        label: 'Alien',
+        lottieID: '1f603',
+      ),
+      const XenEmoji(
+        'https://fonts.gstatic.com/s/e/notoemoji/latest/1f92e/lottie.json',
+        label: 'Vomit',
         lottieID: '1f603',
       ),
     ];
+
+    final dockController = OverlayPortalController();
 
     return Scaffold(
       body: Center(
         child: ListView(
           children: [
-            for (int i = 0; i < 5; i++)
+            for (final (i, lottie) in lotties.indexed)
               Align(
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
@@ -113,20 +130,19 @@ class _XenEmojifyExampleState extends State<XenEmojifyExample> {
                           ),
                           child: XenEmojify(
                             lottieSource: LottieSources.network,
-                            displayLabel: true,
+                            showLabel: false,
                             xenEmojifyDock: XenEmojifyDock(
-                              dockColor: Colors.amber.withOpacity(0.7),
+                              dockColor: Colors.cyan.withOpacity(0.7),
                               onEmojiSelect: (emoji) {
                                 setState(() => initialEmoji = emoji);
                               },
-                              xenEmojis: lottie,
+                              xenEmojis: lotties,
                             ),
-                            initialEmoji: i == 0 ? null : lottie[i],
-                            // placeholderWidget: const MyEmojifyWidget(),
+                            initialEmoji: i == 0 ? null : lottie,
                           ),
                         ),
                         InkWell(
-                          onTap: () {},
+                          onTap: () => dockController.show(),
                           child: const Padding(
                             padding: EdgeInsets.all(8.0),
                             child: Wrap(