+++ lib/src/animations/_hovered_emoji_animation_mixin.dart
@@ -1,16 +1,15 @@
 // BSD License. Copyright © Kiran Paudel. All rights reserved
 
-import 'package:flutter/animation.dart';
+import 'package:flutter/material.dart';
+import 'package:xen_emojify/src/provider/xen_emojify_provider.dart';
+import 'package:xen_emojify/xen_emojify.dart';
 
-///
-mixin HoveredEmojiAnimationMixin {
-  ///
+part 'package:xen_emojify/src/widgets/xen_emojify_dock.dart';
+
+mixin _HoveredEmojiAnimationMixin {
   late List<AnimationController> zoomControllers;
-
-  ///
   late List<Animation<double>> zoomAnimations;
 
-  ///
   void init(TickerProvider vsync, int emojiCount) {
     zoomControllers = List.generate(
       emojiCount,
@@ -28,7 +27,6 @@ mixin HoveredEmojiAnimationMixin {
     ];
   }
 
-  ///
   void disposeControllers() {
     for (final controller in zoomControllers) {
       controller.dispose();