+++ CHANGELOG.md
@@ -5,8 +5,7 @@
 
 # v1.0.1
 
-* 🦄 Smooth zoom animation.
+* 🦄 Smooth animation.
 * 📖 Adds documentation of LottieSources.
 * ✨ Allows user to add their own OverlayPortalController to show and hide dock.
 * ➕ Adds support for web.
-*