+++ CHANGELOG.md
@@ -2,3 +2,10 @@
 
 * 🎉 xen_emojify initial release.
 
+
+# v1.0.1
+
+* 🦄 Smooth animation.
+* 📖 Adds documentation of LottieSources.
+* ✨ Allows user to add their own OverlayPortalController to show and hide dock.
+* ➕ Adds support for web.